+++ educa100/src/main/java/com/educa100/controller/dto/response/NotaResponse.java
@@ -1,4 +1,4 @@
 package com.educa100.controller.dto.response;
 
-public record NotaResponse(int id_nota) {
+public record NotaResponse(Long id_nota) {
 }
