+++ educa100/src/main/java/com/educa100/infra/exception/NotaNotFoundException.java
@@ -1,7 +1,7 @@
 package com.educa100.infra.exception;
 
 public class NotaNotFoundException extends NotFoundException{
-    public NotaNotFoundException(int id){
+    public NotaNotFoundException(Long id){
         super("Nota com o id " +id +" não encontrado");
     }
 }
