+++ educa100/src/main/java/com/educa100/infra/exception/PapelNotFoundException.java
@@ -1,7 +1,7 @@
 package com.educa100.infra.exception;
 
 public class PapelNotFoundException extends NotFoundException{
-    public PapelNotFoundException(int id){
+    public PapelNotFoundException(Long id){
         super("Papel com o id " +id +" não encontrado");
     }
 }
