+++ educa100/src/main/java/com/educa100/service/MateriaService.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public interface MateriaService {
     public MateriaEntity salvar(MateriaEntity materia);
-    public void atualizar(int id);
-    public void removerPorId(int id);
+    public void atualizar(Long id);
+    public void removerPorId(Long id);
     public List<MateriaEntity> listarTodos();
-    public MateriaEntity listarPorId(int id);
-    public List<MateriaEntity> listarPorIdCurso(int id);
+    public MateriaEntity listarPorId(Long id);
+    public List<MateriaEntity> listarPorIdCurso(Long id);
 
 }
